Fixed explicit connect after connection has disconnected
https://github.com/ably/ably-java/issues/167
says that, after a disconnect caused by a network outage, then waiting
long enough for the connection to go disconnected state and the channels
to be detached, an explicit AblyRealtime.connection.connect() is not
immediately actioned. Instead it does not connect until up to a couple
of minutes later.

It looks like the condition is incorrect in ConnectionManager.connect()
to see if a connection attempt is already underway, so the explicit
connect does not do anything. Instead it waits until the next time it
would have attempted a reconnect anyway.

Fixed.

Also fixed so that the connect happens even if there is a connection
thread already started.