RTN23a: implemented heartbeat monitoring in websocket
If we do not see any activity on the websocket transport for
maxIdleInterval+realtimeRequestTimeout, close the websocket so the
transport is disconnected.

maxIdleInterval is in connectionDetails, or defaults to 20s if not
supplied. If it is set to 0, then no heartbeat monitoring is done.

realtimeRequestTimeout defaults to 10s and is not configurable. Some of
the other default timeouts in transport.Defaults probably need to be
changed to realtimeRequestTimeout.