Fixed message serial out of sync after recover
https://github.com/ably/ably-java/issues/170
had a problem where reeonnection from suspended and explicit channel
recovery stops CompletionListener.onSuccess() being called for messages
published after that.

The problem was that the message serial in ConnectionManager and
PendingMessageQueue got out of sync -- only the former was reset back to
0, so then PendingMessageQueue got confused by the message serial in an
ack after a publish, and the callback was not called.

Fixed by ensuring that msgSerial only gets reset when the connection id
changes on connect, and then the PendingMessageQueue is reset at the
same time. Also any pending messages (ones that have been sent but not
acked) at that time are failed.