Fixed message serial out of sync after reconnection
https://github.com/ably/ably-java/issues/170
had a problem where connection recovery and explicit channel
re-attachment stops CompletionListener.onSuccess() being called for
messages published after that.

The problem was that the message serial in ConnectionManager and
PendingMessageQueue got out of sync -- only the former was reset back to
0, so then PendingMessageQueue got confused by the message serial in an
ack after a publish, and the callback was not called.

Fixed by ensuring that both are reset to 0 at the same time.