Separate java build
This introduces a "java" subproject for the JRE-specific version of the
lib, leaving the "lib" subproject containing the common code.

The normal jar also contains the common lib classes from the lib
subproject, without any other dependencies.  There continues to be a
fullJar task to generate a jar containing all dependencies. The Maven
upload is the former, normal, jar, with dependencies set appropriately.

The Maven upload no longer contains javadoc or source classes. If it
did, the jars would be pretty much empty, and I did not spend much time
trying to figure out how to get the javadoc or source from the common
code into the jars here.