ConnectionManager: fixed problems from abandoned transports
I was getting problems where a transport that ConnectionManager had
abandoned due to timing out was, a bit later, generating an
onTransportUnavailable that was not ignored.

Fixed by setting ConnectionManager.transport in connectImpl so we can
tell whether events are from the right transport.

This fix leads to double disconnected notifications from the transport.
So I also needed to fix handleStateChange to stop a transition from
suspended to disconnected.

Also added a 1s delay in RealtimeConnectFailTest.connect_fail_suspended
to ensure that it detects the double disconnected bug.