ConnectionManager: determine fallback earlier
Previously, it was determining whether to try and fall back, and thus
stay in "connecting" state, in checkSuspend, but not determining the
fallback host until connectImpl. The problem with that is that failure
to get a fallback host (the list has been exhausted, or the original
host is non-default so no fallback is done) resulted in "failure" state.

Now it determines the fallback host in checkSuspend, so it is in a
position to change to "disconnected" state if that fails.

This fix also ensures that lack of internet connectivity leads to
"disconnected" rather than "failed".