Http,AsyncHttp: rest fallback should fail with 50x error
When rest fallback failed due to reaching httpMaxRetryCount hosts tried,
or running out of fallback hosts, it was returning a 404 error. This
commit fixes it to return whatever the last failure returned, which must
be a 50x otherwise it would not have been considering fallback.

This matches a change I made recently in realtime fallback to return a
50x on running out of fallbacks.

Also fixed a couple of tests in HttpTests that were testing the old
behaviour.