Separate android build
The previous commit renamed the "android-test" subproject to "android".
This commit modifies it to generate the Android-specific version of the
lib, and run tests on it.

The resulting AAR also contains the common lib classes from the lib
subproject.  The Maven upload has dependencies set appropriately.

The Maven upload does not contain javadoc or source classes. If it did,
the jars would be pretty much empty, and I did not spend much time
trying to figure out how to get the javadoc or source from the common
code into the jars here.

There are some extra test failures over and above what you get on the
JRE-specific library. I have not investigated these yet.