Don't modify ClientOptions
The library was modifying ClientOptions.restHost and/or
ClientOptions.realtimeHost to reflect what it defaulted to, after adding
the environment prefix.

This was causing problems with tests that were using the same
ClientOptions for two library initializations. The second one gave an
error because both the host and environment were set by then.

Fixed.