Enqueue concurrent CalledActivate from WaitingForRegistrationSync.
This isn't in the spec (yet). The previous behavior causes the
activation callback to be called right away with no error, while a
previous activation is ongoing. The correct behavior is to serialize
the calls, so that the second call is processed after the first has
finished.

At the same time, remove the CalledActivate enqueue from
NotActivated, which relied on the old behavior but wasn't doing
anything really useful. (This needs to be removed from the spec
too.)