Add checks to detect attempts to operate a channel cipher from more than one thread.
I could have gone down the route of simplistic addition of locks to the encrypt and decrypt methods, but that would just have been a sticky plaster, probably hiding deeper routed architectural issues elsewhere in this codebase.
By adding this lightweight check, we will see specific evidence that this particular problem is the fault if customers encounter this.