Add messages to queues and remove explicit send call
When sending messages, the reattempted messages from pending queue would be readded because sendImpl will push a new message to the queue if ack is required.
This change clones previous messages from pending queue, using it as a reference to reattempt, but clears queue before reattempt.

So now how it works:
If the connection resumed successfully, pending messages (if available) will be added in front of queued messages, resetting the start serial to the first serial of pending messages.

If connection resume is not successful, msgSerial and start serial will be reset to 0.

This will also remove the explicit call to send() as the send should happen when the connection transition event arrives which should happen after this.

Please note that I kept msgSerial = 0 as this will be increased in sendImpl