Do not supersede pending presence message whilst waiting on attach
At the moment, Presence.pendingPresence is a HashMap by client id. This means that if multiple presence messages are added prior to the channel attaching, all previous messages are superseded by the last.

This is causing issues in ably-asset-tracking-android as the ENTER message for which we are listening is being superseded and therefore our listener never gets called.

This change makes pendingPresence a simple list and so changes the behaviour that all presence messages will be sent upon attaching, without superseding.