Drop messages where channel does not exist
At the moment, if a message is received and a channel does not exist
the channel is created (via channels.get) to process the message.

This can lead to issues with channels.release, which also detaches the channel
after removing it from the channel map. If the timing is unfortunate,
then the DETACH message will come back and re-create the channel, meaning
it cannot be recreated afresh later.

This change fixes the issues by dropping any messages that are received for
a non existent channel and logs the fact that this has happened.

Fixes #913