Make ParameterizedTest an abstract class
Some test runners were picking it up as a test in its own right,
which was then causing JUnit ti complain as it contains no test
case methods.

By making it abstract, it allows us to extend it for protocol testing
but not make it runnable in its own right.