Suspend timer is set when transport is unavailable and last state was connected
Previously, a connection that had been connected for longer than the suspend timer period (set in onConnected)
would have immediately transitioned to suspended when the transport became unavailable. This is
because the timer is not reset the first time the transport fails.

This change fixes this by resetting the suspend timer when the transport becomes unavailable, if
the current connection state is connected.

Fixes #925