Pass uri without query string
authUri was passed including query string, but somewhere else those query strings are already built and passed to the same method. This was causing a malformed authUrl which was causing connection requests with authUrl as part of query of url to fail.

Making this change helped tests I previously wrote for this to pass.