fix: transport not disconnecting after ttl passed
Re-organises the activity timer logic to make sure that the timer runs. Previously the timer existence check in checkActivity() would cancel itself out because it would check (once the timer expires) that the timer was not set, when it was. So the disconnect logic would never get to run and a dead transport would linger until the underlying transport threw an exception which could be up to 2 minutes.

This change fixes this by moving the timer checking / timeout logic away from the existence check in checkActivity(). The arrangement of logic is more in line with the implementation in ably-js.

Fixes #932