test: fix flakey reject_invalid_message_data test
The test was failing as it was setting the log handler and level, which is static and therefore global. This logger is shared between tests that are not running
in isolation, so other tests could write to the log, causing the log assertion to fail.

This change fixes the fails by removing the log assertions and instead asserting on the state of the message post-encoding.

Fixes #946