test: fix resend pending messages flakey test
The test uses a sleep to ensure that the TTLs pass and that the connection is marked as stale. However, it does this whilst the channel is still active, and websocket ping/pongs could still be sent, therefore its possible that on the channel reconnecting, a resume will take place.

This change fixes the bug by performing the wait when the channel has been disconnected (with retries suppressed) which guarantees that the connection will be stale when the test reconnects.

Fixes #942