test: fix attach_when_channel_in_detaching_state
The test was failing/flakey because the channel sometimes quickly cycled
through detaching into attaching, before the listener had a chance to check
the state.

This change fixes the issue by making the assertion via a channel state listener
instead of a point-in-time check of the channel state, which means transient states
are also recorded.

Fixes #944