fix: fallback hosts always being used on transport error
Prior to this change, a websocket transport error (whether that's a clean close,
abnormal close or some others) would immediately result in a fallback host being used,
even if the original host was entirely operational.

This is because the decision on whether to fallback is partly based on the
presence of a "pending connection", which is set when the connecting state is entered. This
pending connection is never cleared, except for when the decision is not to use a fallback. The result
of this, is that a websocket error event (following a successful connection) will always have a lingering pending
connection, so if the error from the websocket is "server-related", it will automatically use a fallback host.

This change fixes the issue by clearing the pending connection whenever a connected state is reached. By doing this,
it means that a simple disconnection (e.g. from a scaling event) will automatically retry the same host, wheras a disconnection
during the connection process will try a fallback.

The behaviour implemented is in the spirit of RTN17c, but the spec could do with an update to make it clearer as to what lifecycle conditions in a Realtime connection constitutes the need to use a fallback host (our interpretation is that the spec references RSC17l, which refers to errors in the REST API, so it follows that we'd expect this to apply to the CONNECTING state for a Realtime connection.

Fixes #950