fix: use appropriate error code for channel attachment timeout
Previously, for channel attachment timeouts we were using error code 91200 which is not documented for users and was removed from the specification somewhere
around version 0.9. It looks like it was removed almost immediately after it was implemented in ably-java.

This change fixes this by changing the code to 90007 which is used elsewhere
and more appropriate for this use-case (used in detach in ably-java, and in ably-js).

Fixes #959